source-mysql: Move binlog expiry sanity-check into prerequisites
Previously this was done as part of the connect operation, but
this is the more appropriate place for the check to be done.
